+++ fawkes/algorithms/sentiment.py
