+++ fawkes/constants.py
