+++ fawkes/email_summary/templates/weekly-detailed-review-block.html
