+++ fawkes/email_summary/templates/weekly-detailed.html
