+++ fawkes/scripts/fawkes-secrets.env
@@ -1,0 +1,5 @@
+export FAWKES_ENV_SENDGRID_API_KEY="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"
+export FAWKES_ENV_TWITTER_CONSUMER_KEY="DSyjE5HnUcry5njxss1z1Qjqs"
+export FAWKES_ENV_TWITTER_CONSUMER_SECRET="BQwBxdHuU3r47CLhqZSgyQRgdY4jLX0QZgpWv7jGlohPUFINLV"
+export FAWKES_ENV_TWITTER_ACCESS_TOKEN_KEY="874195199748063233-Tb9KNIMN389UIGfhAsi0uEd6bQR0nO1"
+export FAWKES_ENV_TWITTER_ACCESS_TOKEN_SECRET="pPwQQdimZpi7CXNsxydHBXSqu25eeVJJ49H08egfGNgZ5"
