+++ fawkes/scripts/install-elastic.sh
