+++ fawkes/scripts/run-all.sh
